requirements.txt
# Charles_FocusedSpec - Requirements
# מערכת חיזוי מניות פורצות

# Core Data Science & Analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Financial Data & APIs
yfinance>=0.2.0
requests>=2.25.0
beautifulsoup4>=4.11.0
curl_cffi>=0.9.0

# Technical Analysis
pandas_ta>=0.3.14b0
ta>=0.11.0

# Visualization & Dashboard
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.6.0
altair>=5.0.0

# Machine Learning & AI
transformers>=4.46.0
torch>=1.13.0
tensorflow>=2.10.0
scikit-learn>=1.0.0

# Natural Language Processing
nltk>=3.8
spacy>=3.5.0
textblob>=0.17.0

# Database & Storage
sqlite3
peewee>=3.18.0
chromadb>=0.4.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.0
configparser

# Utilities
tqdm>=4.67.0
joblib>=1.4.0
psutil>=5.9.0
platformdirs>=4.3.0

# Development & Testing
pytest>=8.3.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Web & Networking
websockets>=13.1
aiohttp>=3.8.0
asyncio

# Date & Time
python-dateutil>=2.8.0
pytz>=2022.6

# File Processing
openpyxl>=3.1.0
xlrd>=2.0.0
gzip
pathlib

# Security & Cryptography
cryptography>=45.0.0
certifi>=2025.7.0

# Git Integration
gitpython>=3.1.0

# JSON & Data Validation
jsonschema>=4.25.0
jsonschema_specifications>=2025.4.0

# Monitoring & Logging
watchdog>=6.0.0
blinker>=1.9.0

# Additional Financial Tools
finnhub-python>=2.4.0
alpha-vantage>=2.3.0
quandl>=3.7.0

# Optional: Advanced Features
# Uncomment if needed:
# tensorflow-gpu>=2.10.0
# torchvision>=0.14.0
# opencv-python>=4.7.0
# selenium>=4.8.0 